+++ react/game/src/app/screens/Game/components/Square/index.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class Square extends React.Component {
+  render() {
+    return (
+      <button className="square">
+        {/* TODO */}
+      </button>
+    );
+  }
+}
+
+export default Square;
